Check for CSRF token

+++ public/actions/deleteComment.php
@@ -6,12 +6,17 @@ require '../controllers/commentsController.php';
 
 $controller = new CommentsController();
 
-$comment = $controller->deleteComment(htmlspecialchars($_GET["id"]));
+if ($_GET['csrf'] === $_SESSION['CSRF-Delete']) {
+    $_SESSION['CSRF-Delete'] = NULL;
+    $comment = $controller->deleteComment(htmlspecialchars($_GET["id"]));
 
-if ($comment === -1) {
-    header("Location: /index.php?error=delete");
-    header("Location: /index.php?success=delete");
+    if ($comment === -1) {
+        header("Location: /index.php?error=delete");
+    }
+    else {
+        header("Location: /index.php?success=delete");
+    }
 }
 else {
-    header("Location: /index.php?success=delete");
+    eader("Location: /index.php?error=delete");
 }
